# transportation-api/docs/swagger.yaml
components:
  schemas:
    entity.Airport:
      properties:
        iata:
          type: string
        location:
          $ref: '#/components/schemas/entity.Location'
        municipality:
          type: string
        name:
          type: string
      required:
      - iata
      - location
      - municipality
      - name
      type: object
    entity.AmbiguousFlightChoice:
      properties:
        departureDateTime:
          type: string
        destinationIata:
          type: string
        originIata:
          type: string
      required:
      - departureDateTime
      - destinationIata
      - originIata
      type: object
    entity.ErrAmbiguousFlightRequest:
      additionalProperties:
        items:
          $ref: '#/components/schemas/entity.AmbiguousFlightChoice'
        type: array
      type: object
    entity.Flight:
      properties:
        geoJson:
          type: object
        legs:
          items:
            $ref: '#/components/schemas/entity.FlightLeg'
          type: array
          uniqueItems: false
      required:
      - geoJson
      - legs
      type: object
    entity.FlightLeg:
      properties:
        aircraft:
          nullable: true
          type: string
        airline:
          type: string
        amadeusFlightDate:
          nullable: true
          type: string
        arrivalDateTime:
          type: string
        departureDateTime:
          type: string
        destination:
          $ref: '#/components/schemas/entity.Airport'
        durationInMinutes:
          type: integer
        flightNumber:
          type: string
        origin:
          $ref: '#/components/schemas/entity.Airport'
      required:
      - aircraft
      - airline
      - amadeusFlightDate
      - arrivalDateTime
      - departureDateTime
      - destination
      - durationInMinutes
      - flightNumber
      - origin
      type: object
    entity.Location:
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
      - latitude
      - longitude
      type: object
...
    request.Flight:
      properties:
        legs:
          items:
            $ref: '#/components/schemas/request.FlightLeg'
          type: array
          uniqueItems: false
      required:
      - legs
      type: object
    request.FlightLeg:
      properties:
        date:
          example: "2026-01-30"
          type: string
        flightNumber:
          example: EK412
          type: string
        originAirport:
          example: SYD
          nullable: true
          type: string
      required:
      - date
      - flightNumber
      - originAirport
      type: object
...
    response.Error:
      properties:
        detail:
          nullable: true
          type: string
        error:
          type: string
      required:
      - error
      type: object
...
info:
  title: Kompass Transportation API
  version: "1.0"
openapi: 3.1.0
paths:
  /flights:
    post:
      operationId: postFlight
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request.Flight'
        description: flight
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.Flight'
          description: OK
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.ErrAmbiguousFlightRequest'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.Error'
          description: Internal Server Error
      summary: Find flight
      tags:
      - flights
...
servers:
- url: http://127.0.0.1:8080/api/v1